Fix conversion warnings in debugger
Conversion warnings fixed for following commands:

* loader.function.map.show
* loader.block.map.show

Address types changed from unsigned to uint64_t.

+++ src/front/wdb.cpp
@@ -758,7 +758,7 @@ bool command_dispatch(string& command, vector<string>& operands, CPU& cpu, State
         cout << "  * return value:    " << top->place_return_value_in << endl;
         cout << "  * resolve return:  " << (top->resolve_return_value_register ? "yes" : "no") << endl;
     } else if (command == "loader.function.map.show") {
-        unsigned addr;
+        uint64_t addr;
         bool exists = false;
         for (string fun : operands) {
             try {
@@ -775,7 +775,7 @@ bool command_dispatch(string& command, vector<string>& operands, CPU& cpu, State
             }
         }
     } else if (command == "loader.block.map.show") {
-        unsigned addr;
+        uint64_t addr;
         bool exists = false;
         for (string fun : operands) {
             try {
